% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMCE fx for CRAN.R
\name{plotAMCE}
\alias{plotAMCE}
\title{plotAMCE: Beautiful AMCE plots}
\usage{
plotAMCE(
  Model,
  Type = "Nothing",
  version = NULL,
  trait_renames = NULL,
  attribute_renames = NULL,
  trait_levels = NULL,
  plot_title = NULL,
  plot_subtitle = "Average Marginal Component Effects",
  y_label = expression(Delta * " Predicted Probability of Selection (Percentage)")
)
}
\arguments{
\item{Model}{An AMCE object (created by amce() from cjoint package)}

\item{version}{Choose wide plots or tall plots (when version = "tall"). Defaults to wide.}

\item{trait_renames}{Optional field. Used to rename list of within-attribute traits (levels) from original dataframe for presentation.}

\item{attribute_renames}{Optional field. Used to rename list of attributes from original dataframe for presentation.}

\item{trait_levels}{Optional field listing within-attribute traits (levels).}

\item{plot_title}{Title of plot}

\item{plot_subtitle}{Subtitle of plot}
}
\description{
Plot average marginal component effects (AMCEs) for displaying conjoint results
}
\examples{
#'
library(cjoint)
data(immigration)
model <- amce(Chosen_Immigrant ~ Gender + Country + Education, data = immigration,
              respondent.id = "ResID")
plotAMCE(model)
}
