% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mms fx.R
\name{plotMM}
\alias{plotMM}
\title{plotMM: Beautiful Marginal Mean plots}
\usage{
plotMM(
  Model,
  Type = "Nothing",
  version = NULL,
  trait_renames = NULL,
  attribute_renames = NULL,
  trait_levels = NULL,
  plot_title = NULL,
  plot_subtitle = "Marginal Means",
  y_label = expression("Average Probability of Choosing Profile")
)
}
\arguments{
\item{Model}{A marginal mean object (created by mm() from cregg package (https://github.com/leeper/cregg))}

\item{version}{Choose wide plots or tall plots (when version = "tall"). Defaults to wide.}

\item{trait_renames}{Optional field. Used to rename list of within-attribute traits (levels) from original dataframe for presentation.}

\item{attribute_renames}{Optional field. Used to rename list of attributes from original dataframe for presentation.}

\item{trait_levels}{Optional field listing within-attribute traits (levels).}

\item{plot_title}{Title of plot}

\item{plot_subtitle}{Subtitle of plot}
}
\description{
Plot marginal means for displaying conjoint results
}
\examples{

library(cjoint)
library(cregg)
data(immigration)
model <- amce(Chosen_Immigrant ~ Gender + Country + Education, data = immigration,
              respondent.id = "ResID")
model1 <- mm(model, Chosen_Immigrant ~ Gender + Country + Education, id=~"ResID"
plotMM(model1)
}
